# Fix jax's `in_top_k` and numpy's `top_k`