Fix #19995

The performance will be slightly slower, but it should be acceptable.

|Before|After||
|-|-|-|
|10.842s|11.063s|+2%|

Benchmark script:

```python
import functools
import timeit

import jax

from keras.src.backend.jax.math import in_top_k

logits = jax.random.uniform(jax.random.key(42), shape=(32, 128, 50))
labels = jax.random.randint(
    jax.random.key(42), shape=(32, 128), minval=0, maxval=50
)


@functools.partial(jax.jit, static_argnames=["k"])
def jit_in_top_k(logits, labels, k=5):
    return in_top_k(labels, logits, k=k)


print(jit_in_top_k(logits, labels, k=5).mean())
print(timeit.timeit(lambda: jit_in_top_k(logits, labels, k=5)))

```

Additionally, improve numpy's `top_k` which had significantly poor performance due to double sorting (or even more times of sorting...)